resolve issues found in PR review
